deps: update required dependency versions
The gem does not actually depend on i18n being <1. There is a PR #33 that just changes the version. This suggests that the code works with i18n >1.
The gemspec should allow that.

Also, older country translations is also supported. They were upgraded in #30 because of security concerns, without any code changes. Bundler warns and advises for security. This gem should not force people to upgrade.